Python/requirements.txt
pymongo==3.11.2
Flask==1.1.2
flask-cors==3.0.9
beem==0.24.26
spacy==3
nmslib==2.1.1
ftfy==5.8
numpy==1.21.0
fasttext==0.9.2
aiohttp==3.7.4
nltk==3.5
waitress==1.4.4
markdown==3.3.3
fastapi==0.63.0
uvicorn==0.13.3
motor==2.3.1
bs4==0.0.1
hivesigner==0.1.3
torch==1.9.0
torchvision
gensim==3.8.3
python-dotenv==0.15.0
streamlit==0.88.0